chore(tests): refactor and separate out RRset test base class
